backend/requirements.txt
Flask==2.2.5
Flask-CORS==4.0.0
ultralytics==8.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0